+++ client/package.json
@@ -73,6 +73,7 @@
         "eslint-plugin-jsx-a11y": "^6.8.0",
         "eslint-plugin-react": "^7.33.2",
         "eslint-plugin-react-hooks": "^5.1.0",
+        "http-status-codes": "^2.3.0",
         "jest-fetch-mock": "^3.0.3",
         "katex": "^0.16.11",
         "markdown-it-katex": "^2.0.3",
