+++ client/test/integration/Auth/WaitAndNavigate.test.tsx
@@ -1,5 +1,5 @@
 import { act, screen } from '@testing-library/react';
-import { getValidationResults } from 'route/auth/VerifyConfig';
+import { getValidationResults } from 'util/config'; // Globally mocked
 import { mockAuthState } from 'test/__mocks__/global_mocks';
 import { setupIntegrationTest } from 'test/integration/integration.testUtil';
 
@@ -8,29 +8,29 @@ jest.useFakeTimers();
 const authStateWithError = { ...mockAuthState, error: Error('Test Error') };
 const setup = () => setupIntegrationTest('/library', authStateWithError);
 Object.defineProperty(window, 'location', {
-  value: {
-    ...window.location,
-    reload: jest.fn(),
-  },
-  writable: true,
+    value: {
+        ...window.location,
+        reload: jest.fn(),
+    },
+    writable: true,
 });
 
 describe('WaitAndNavigate', () => {
-  beforeEach(async () => {
-    (getValidationResults as jest.Mock).mockReturnValue(
-      Promise.resolve({ configField: 'test' }),
-    );
-    await setup();
-  });
-
-  it('redirects to the WaitAndNavigate page when getting useAuth throws an error', async () => {
-    expect(screen.getByText('Oops... Test Error')).toBeVisible();
-    expect(screen.getByText('Waiting for 5 seconds...')).toBeVisible();
-
-    await act(async () => {
-      jest.advanceTimersByTime(5000);
+    beforeEach(async () => {
+        (getValidationResults as jest.Mock).mockReturnValue(
+            Promise.resolve({}),
+        );
+        await setup();
     });
 
-    expect(screen.getByText(/Sign In with GitLab/i)).toBeVisible();
-  });
+    it('redirects to the WaitAndNavigate page when getting useAuth throws an error', async () => {
+        expect(screen.getByText('Oops... Test Error')).toBeVisible();
+        expect(screen.getByText('Waiting for 5 seconds...')).toBeVisible();
+
+        await act(async () => {
+            jest.advanceTimersByTime(5000);
+        });
+
+        expect(screen.getByText(/Sign In with GitLab/i)).toBeVisible();
+    });
 });
