+++ client/test/integration/Routes/Signin.test.tsx
@@ -1,22 +1,22 @@
 import { screen, act } from '@testing-library/react';
 import { setupIntegrationTest } from 'test/integration/integration.testUtil';
-import { getValidationResults } from 'route/auth/VerifyConfig';
+import { getValidationResults } from 'util/config'; // Globally mocked
 import { testPublicLayout } from './routes.testUtil';
 
 const setup = () => setupIntegrationTest('/');
 
 describe('Signin', () => {
-  it('renders the Sign in page with the Public Layout correctly', async () => {
-    (getValidationResults as jest.Mock).mockReturnValue(
-      Promise.resolve({ configField: 'test' }),
-    );
-    await act(async () => {
-      await setup();
+    it('renders the Sign in page with the Public Layout correctly', async () => {
+        (getValidationResults as jest.Mock).mockReturnValue(
+            Promise.resolve({}),
+        );
+        await act(async () => {
+            await setup();
+        });
+        await testPublicLayout();
+        expect(
+            screen.getByRole('button', { name: /Sign In with GitLab/i }),
+        ).toBeVisible();
+        expect(screen.getByTestId(/LockOutlinedIcon/i)).toBeVisible();
     });
-    await testPublicLayout();
-    expect(
-      screen.getByRole('button', { name: /Sign In with GitLab/i }),
-    ).toBeVisible();
-    expect(screen.getByTestId(/LockOutlinedIcon/i)).toBeVisible();
-  });
 });
